Refactored to combine Opaque code with the other remote desktop client code.

+++ Opaque/src/main/java/com/undatech/opaque/ConnectionSetupActivity.java
@@ -367,10 +367,8 @@ public class ConnectionSetupActivity extends Activity {
     @Override
     public boolean onOptionsItemSelected(MenuItem menuItem) {
         int itemID = menuItem.getItemId();
-        switch (itemID) {
-        case R.id.actionDeleteConnection:
+        if (itemID == R.id.actionDeleteConnection) {
             deleteConnection (menuItem);
-            break;
         }
         return true;
     }
